+++ common-contract-enricher/src/main/java/com/dell/cpsd/common/contract/enricher/plugins/jsonschema2pojo/MessageAnnotator.java
@@ -5,16 +5,18 @@
 
 package com.dell.cpsd.common.contract.enricher.plugins.jsonschema2pojo;
 
-import com.dell.cpsd.common.rabbitmq.annotation.Message;
-import com.dell.cpsd.common.rabbitmq.annotation.MessageContentType;
-import com.dell.cpsd.common.rabbitmq.annotation.stereotypes.MessageError;
-import com.dell.cpsd.common.rabbitmq.annotation.stereotypes.MessageEvent;
-import com.dell.cpsd.common.rabbitmq.annotation.stereotypes.MessageReply;
-import com.dell.cpsd.common.rabbitmq.annotation.stereotypes.MessageRequest;
+import org.jsonschema2pojo.AbstractAnnotator;
+
+import com.dell.cpsd.contract.extension.amqp.annotation.Message;
+import com.dell.cpsd.contract.extension.amqp.annotation.MessageContentType;
+import com.dell.cpsd.contract.extension.amqp.annotation.stereotypes.ErrorMessage;
+import com.dell.cpsd.contract.extension.amqp.annotation.stereotypes.EventMessage;
+import com.dell.cpsd.contract.extension.amqp.annotation.stereotypes.ReplyMessage;
+import com.dell.cpsd.contract.extension.amqp.annotation.stereotypes.RequestMessage;
+import com.dell.cpsd.contract.extension.amqp.annotation.stereotypes.StereotypeMessage;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.sun.codemodel.JAnnotationUse;
 import com.sun.codemodel.JDefinedClass;
-import org.jsonschema2pojo.AbstractAnnotator;
 
 /**
  * <p>
@@ -83,26 +85,26 @@ public class MessageAnnotator extends AbstractAnnotator
         if (stereotype != null)
         {
             String stereoTypeValue = stereotype.asText();
-            if ("REQUEST".equalsIgnoreCase(stereoTypeValue))
+            if (StereotypeMessage.REQUEST.toString().equalsIgnoreCase(stereoTypeValue))
             {
-                JAnnotationUse annotation = clazz.annotate(MessageRequest.class);
+                JAnnotationUse annotation = clazz.annotate(RequestMessage.class);
                 JsonNode replyToProperty = meta.get("replyToProperty");
                 if (replyToProperty != null)
                 {
                     annotation.param("replyToProperty", replyToProperty.asText());
                 }
             }
-            else if ("REPLY".equalsIgnoreCase(stereoTypeValue))
+            else if (StereotypeMessage.REPLY.toString().equalsIgnoreCase(stereoTypeValue))
             {
-                clazz.annotate(MessageReply.class);
+                clazz.annotate(ReplyMessage.class);
             }
-            else if ("EVENT".equalsIgnoreCase(stereoTypeValue))
+            else if (StereotypeMessage.EVENT.toString().equalsIgnoreCase(stereoTypeValue))
             {
-                clazz.annotate(MessageEvent.class);
+                clazz.annotate(EventMessage.class);
             }
-            else if ("ERROR".equalsIgnoreCase(stereoTypeValue))
+            else if (StereotypeMessage.ERROR.toString().equalsIgnoreCase(stereoTypeValue))
             {
-                clazz.annotate(MessageError.class);
+                clazz.annotate(ErrorMessage.class);
             }
         }
     }