+++ common-contract-enricher/src/main/java/com/dell/cpsd/common/contract/enricher/plugins/jsonschema2pojo/rule/DefaultPropertiesRule.java
@@ -28,10 +28,10 @@ import static java.util.Arrays.asList;
 public class DefaultPropertiesRule extends PropertiesRule
 {
     protected List<ClassAction> actions = asList(new AddInterface(asList("correlationId", "replyTo", "timestamp"),
-                    "com.dell.cpsd.common.rabbitmq.message.MessagePropertiesContainer"),
-            new AddInterface(asList("code", "message"), "com.dell.cpsd.common.rabbitmq.message.ErrorContainer"),
-            new AddGenericInterface("messageProperties", "com.dell.cpsd.common.rabbitmq.message.HasMessageProperties"),
-            new AddGenericInterface("errors", "com.dell.cpsd.common.rabbitmq.message.HasErrors").unwrapFieldType(List.class));
+                    "com.dell.cpsd.contract.extension.amqp.message.MessagePropertiesContainer"),
+            new AddInterface(asList("code", "message"), "com.dell.cpsd.contract.extension.amqp.message.ErrorContainer"),
+            new AddGenericInterface("messageProperties", "com.dell.cpsd.contract.extension.amqp.message.HasMessageProperties"),
+            new AddGenericInterface("errors", "com.dell.cpsd.contract.extension.amqp.message.HasErrors").unwrapFieldType(List.class));
 
     public DefaultPropertiesRule(RuleFactory ruleFactory)
     {