+++ common-rabbitmq/src/main/java/com/dell/cpsd/common/rabbitmq/message/DefaultMessageConverterFactory.java
@@ -5,14 +5,15 @@
 
 package com.dell.cpsd.common.rabbitmq.message;
 
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
+import java.text.SimpleDateFormat;
+
 import org.springframework.amqp.support.converter.ClassMapper;
 import org.springframework.amqp.support.converter.Jackson2JsonMessageConverter;
 import org.springframework.amqp.support.converter.MessageConverter;
 
-import java.text.SimpleDateFormat;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
 
 /**
  * The is a factory for a default message converter configuration.