+++ common-rabbitmq/src/test/java/com/dell/cpsd/common/rabbitmq/MessageAnnotationProcessorTest.java
@@ -1,8 +1,10 @@
 /**
- * Copyright &copy; 2017 Dell Inc. or its subsidiaries.  All Rights Reserved.
- * Dell EMC Confidential/Proprietary Information
+ * Copyright &copy; 2017 Dell Inc. or its subsidiaries. All Rights Reserved. Dell EMC Confidential/Proprietary Information
  */
+
 package com.dell.cpsd.common.rabbitmq;
+
+import static org.junit.Assert.assertEquals;
 
 import java.lang.annotation.Annotation;
 import java.util.Arrays;
@@ -11,14 +13,14 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-import com.dell.cpsd.common.rabbitmq.annotation.Message;
+import com.dell.cpsd.annotation.message.test.data.TestMessageForAutoAnnotationScan;
+import com.dell.cpsd.contract.extension.amqp.annotation.Message;
 
 /**
  * TODO: Document usage. Set proper Vision version in since tag.
  *
  * <p>
- * Copyright &copy; 2017 Dell Inc. or its subsidiaries.  All Rights Reserved.
- * Dell EMC Confidential/Proprietary Information
+ * Copyright &copy; 2017 Dell Inc. or its subsidiaries. All Rights Reserved. Dell EMC Confidential/Proprietary Information
  * </p>
  *
  * @since x.y.z
@@ -33,6 +35,12 @@ public class MessageAnnotationProcessorTest
     {
         callback = Mockito.mock(MessageAnnotationProcessorCallback.class);
         processor = new MessageAnnotationProcessor();
+    }
+
+    @Test
+    public void testScanAndProcessClassMappings()
+    {
+
     }
 
     @Test
@@ -76,6 +84,35 @@ public class MessageAnnotationProcessorTest
         processor.process(callback, clazz);
 
         Mockito.verify(callback, Mockito.never()).found(Mockito.anyString(), Mockito.any());
+    }
+
+    @Test
+    public void testScanAndProcessAnnotation()
+    {
+        processor.scanAndProcessAnnotation(new MessageAnnotationProcessorCallback()
+        {
+
+            @Override
+            public void found(String messageType, Class messageClass)
+            {
+                assertEquals("test.message.request", messageType);
+                assertEquals(TestMessageForAutoAnnotationScan.class, messageClass);
+            }
+        }, "com.dell.cpsd.annotation.message.test.data");
+    }
+
+    @Test
+    public void testScanAndProcessAnnotationNoMatches()
+    {
+        processor.scanAndProcessAnnotation(callback, "com.dell.cpsd.annotation.message.test.data.no.package");
+        Mockito.verify(callback, Mockito.never()).found(Mockito.anyString(), Mockito.any());
+    }
+
+    @Test
+    public void testScanAndProcessAnnotationDefaultPackage()
+    {
+        processor.scanAndProcessAnnotation(callback);
+        Mockito.verify(callback, Mockito.atLeastOnce()).found(Mockito.anyString(), Mockito.any());
     }
 
 }