+++ common-rabbitmq/src/test/java/com/dell/cpsd/common/rabbitmq/TestRequestMessage.java
@@ -5,8 +5,8 @@
 
 package com.dell.cpsd.common.rabbitmq;
 
-import com.dell.cpsd.common.rabbitmq.annotation.Message;
-import com.dell.cpsd.common.rabbitmq.message.HasMessageProperties;
+import com.dell.cpsd.contract.extension.amqp.annotation.Message;
+import com.dell.cpsd.contract.extension.amqp.message.HasMessageProperties;
 
 /**
  * <p>